requirements.txt
numpy
pandas
joblib
pyyaml
pycountry_convert
sklearn